requirements.txt
absl-py==2.3.0
alembic==1.16.1
astunparse==1.6.3
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
Flask==3.1.1
Flask-Migrate==4.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
gast==0.6.0
git-filter-repo==2.47.0
google-pasta==0.2.0
grpcio==1.73.0
h5py==3.14.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
keras==3.10.0
lazy_loader==0.4
libclang==18.1.1
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.1.0
networkx==3.2.1
numpy==2.0.2
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pillow==11.2.1
protobuf==5.29.5
Pygments==2.19.1
PyMySQL==1.1.1
python-dotenv==1.1.0
requests==2.32.4
rich==14.0.0
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
SQLAlchemy==2.0.41
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2024.8.30
tomli==2.2.1
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.22.0
